+++ modules/dxp/apps/commerce/commerce-address-api/src/main/java/com/liferay/commerce/address/service/CommerceRegionLocalServiceUtil.java
@@ -55,7 +55,7 @@ public class CommerceRegionLocalServiceUtil {
 
 	public static com.liferay.commerce.address.model.CommerceRegion addCommerceRegion(
 		long commerceCountryId, java.lang.String name,
-		java.lang.String abbreviation, int priority, boolean published,
+		java.lang.String abbreviation, double priority, boolean published,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService()
@@ -129,7 +129,7 @@ public class CommerceRegionLocalServiceUtil {
 
 	public static com.liferay.commerce.address.model.CommerceRegion updateCommerceRegion(
 		long commerceRegionId, java.lang.String name,
-		java.lang.String abbreviation, int priority, boolean published)
+		java.lang.String abbreviation, double priority, boolean published)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService()
 				   .updateCommerceRegion(commerceRegionId, name, abbreviation,
