+++ modules/dxp/apps/commerce/commerce-address-api/src/main/java/com/liferay/commerce/address/service/CommerceRegionLocalServiceWrapper.java
@@ -49,7 +49,7 @@ public class CommerceRegionLocalServiceWrapper
 	@Override
 	public com.liferay.commerce.address.model.CommerceRegion addCommerceRegion(
 		long commerceCountryId, java.lang.String name,
-		java.lang.String abbreviation, int priority, boolean published,
+		java.lang.String abbreviation, double priority, boolean published,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _commerceRegionLocalService.addCommerceRegion(commerceCountryId,
@@ -129,7 +129,7 @@ public class CommerceRegionLocalServiceWrapper
 	@Override
 	public com.liferay.commerce.address.model.CommerceRegion updateCommerceRegion(
 		long commerceRegionId, java.lang.String name,
-		java.lang.String abbreviation, int priority, boolean published)
+		java.lang.String abbreviation, double priority, boolean published)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _commerceRegionLocalService.updateCommerceRegion(commerceRegionId,
 			name, abbreviation, priority, published);
